tests/sandbox: Substitute alias for libressl.nc
Another one of a series of aliases to rename, similar to the last few
commits. Unfortunately, I didn't evaluate the sandbox test to make sure
all the references were changed, so I missed the netcat one.

This time however, I evaluated the sandbox test with latest nixpkgs and
it went without errors.

Signed-off-by: aszlig <aszlig@nix.build>