profiles/{base,packages}: Add interesting tools
These tools are `modern` replacements for the tried and tested, if a bit
boring, *NIX tools that we all love. My motivation is to try these tools
out for an extended period of time and see how I get on with them.

In order to remind myself that a certain tool is installed I've added
some aliases. Let's see how this goes.