machines: Remove slylandro
As mentioned in the initial commit (e73fcff03faed773df2500965cb9c4a4fc),
the machine was only temporary as a substitute for tishtushi.

Since slylandro had a pretty slow dual core CPU and its own quirks, this
was never a long-term solution and for the time being my intentions are
to work with dnyarri's new hardware until I have a less annoying setup
when I'm on the road again.

While writing this message, slylandro just died a gruesome death with
"cryptsetup erase", followed by "blkdiscard" on the whole drive.

Signed-off-by: aszlig <aszlig@nix.build>