profiles/packages: Remove nixops for now
It's not clear when the `certifi` PR[1] will be merged, and I don't want
to have to wait for a new hydra build because of it. `nixopsUnstable` is
unusable in its current from due to almost lacking documentation, and so
I have little choice but to remove `nixops` right now until `certifi` is
merged.

[1]:https://github.com/NixOS/nixpkgs/pull/127453