aszlig/vim: Switch to dedicated plugin for Jinja2
The Vim syntax highlighting plugin file is no longer[1] shipped with
Jinja2 version 3.x, so the build fails accordingly with:

  install: cannot stat 'ext/Vim/jinja.vim': No such file or directory

In another upstream pull request[2] one of the project members mentioned
another syntax plugin which apparently seems to be more up to date. This
is what I'm hereby switching to as a replacement.

[1]: https://github.com/pallets/jinja/pull/1196
[2]: https://github.com/pallets/jinja/issues/1007

Signed-off-by: aszlig <aszlig@nix.build>