vim: Fix Fish syntax file
The file in question actually was a ZIP file, which instead of being
unpacked got directly moved to syntax/fish.vim and in turn caused errors
whenever the filetype was set to "fish".

Instead of just fixing up the ZIP file I switched to a GitHub repository
that seemed to be maintained a lot more (last commit in 2020) than the
one we had so far (last change 2013).

Signed-off-by: aszlig <aszlig@nix.build>