modules/git: Get rid of "with lib;" and refactor
This gets rid of bringing all of the lib attribute into the scope of the
whole module and also refactors a bit of ugliness, for example we're now
using overrideAttrs instead of overrideDerivation.

In addition I've simplified the type for the configuration, which should
now make it possible to mix subsubsections with normal subsections.

Signed-off-by: aszlig <aszlig@nix.build>