modules/git: Switch to RFC-0042
In RFC-0042[1], the structural settings are exposed via a "settings"
attribute, while in our module it's called "config". To make this less
ambiguous (since there is already a "config" attribute passed to
modules) and more in line with best practices, I renamed it accordingly.

Additionally, the configuration file is now generated via *defining*
another (read-only) option, which can be used by other modules to
reference the path. The previous way this has been done was using the
apply attribute to mkOption, which makes it really hard to access the
original attributes for these settings.

[1]: https://git.io/JcXmU

Signed-off-by: aszlig <aszlig@nix.build>