workstation: Advertise XTerm true color capability
So far, the TERM environment variable has been set to xterm-256color,
but in reality newer XTerm versions already supported 24bit colors so
setting this to xterm-direct results in using the right terminfo entry
for our terminal.

To make sure this is really the case, let's explicitly set directColor
to true, because while it is enabled in nixpkgs by default it is however
a compile-time option and could possibly be disabled.

Additionally, Vim is now looking pretty gruesome because my colorscheme
so far has used colors for 16-color terminals and I don't particularly
like the GUI colors. I added a few fixups for the color scheme to
address that.

Signed-off-by: aszlig <aszlig@nix.build>