vim: Replace vim-css-color by hexokinase
So far vim-css-color worked quite well for what I wanted, but after
talking to @devhell about possible alternatives, I stumbled upon
hexokinase and tried it a bit.

One of the gripes I had with things such as colorizer is that it
highlights colors regardless of the file types we're in, which in turn
will also highlight things where the hash character is not a hex value,
for example in Erlang's base notation for integers.

Hexokinase also highlights all file types but first of all, it only
highlights things separated by word boundary and also it's way less
obtrusive because the way I've configured it only the hash character is
highlighted, not the whole color value.

Signed-off-by: aszlig <aszlig@nix.build>