modules/core: Fix order of Vuizvui pkgs namespace
So far, we have passed self instead of super to our custom package set,
but this makes it very hard when overriding other packages and refering
to one of these packages in Vuizvui.

To fix this, I not only used mkBefore to make sure that the overlay
comes before every other overlay but also using super makes sure that
whenever we do overrides in Vuizvui, the package *before* the current
overlay is selected instead of possibly running into an infinite
recursion.

Signed-off-by: aszlig <aszlig@nix.build>