modules/git: Fix reference to Git config file
Regression introduced in 709e45c2a8c231201d3f67c6a954021ca2a5f796.

With that commit I renamed the "config" option to be called "settings",
but not only that, the actual config *file* is now to be found in the
"configFile" option.

Unfortunately I forgot to change that reference, which caused an
evaluation error.

Signed-off-by: aszlig <aszlig@nix.build>