workstation: Switch Git diff viewer to delta
I'm already using this config since a few months and since I'm slowly
getting used to it, let's try to make it the default for my workstation
profile.

Unfortunately, libgit2 uses the Git configuration from /etc, but our
configuration is directly patched into Git and not in /etc but in
"${pkgs.git}/etc", so we need to patch libgit2 to use the right
configuration file.

Another goof is that we can't use ${pkgs.delta} directly in our Git
configuration because it would introduce a circular dependency between
Git itself and delta (which uses libgit2 which in turn refers to
"${pkgs.git}/etc"), so for the time being I'm relying on $PATH for
delta.

Signed-off-by: aszlig <aszlig@nix.build>