pkgs/vim: Fix color for LightBlue in color scheme
If we use 256 color mode in XTerm, using LightBlue in Vim results in
0x5fd7ff but LightBlue in GUI mode will use 0xadd8e6 which has a low
contrast to the default color (0xbebebe).

Since my eyes are not getting better with age, I decided to go with the
old color code that provides better contrast even though I'm quite happy
with the rest of the "more nuanced" colors.

Signed-off-by: aszlig <aszlig@nix.build>