machines/sterneseemann/base: hack around makewhatis(8) ignoring secs
buildEnv is conservative with creating symlinks, i. e. it only creates
directories if it has to. Consequently if a directory is only present in a
single package in the environment, it'll be a symlink.

Enter: makewhatis(8), a tool that has never imagined a creation as accursed as
Nix or even NixOS. Thus it assumes that probably no one ever uses symlinks in
their man directory and if they do, it'll be to alias man pages. Consequently
it assumes that all symbolic links are files [1] and ignores them in the
normal mode because they are in the wrong place. To still be able to use
apropos(1) with POSIX man pages, introduce this shell hack to re-create the
symlinked directories before building mandoc's db.

[1]: See also
https://inbox.vuxu.org/mandoc-tech/bccac2cd-01b6-b349-86e5-de4066ed8dee@systemli.org/T/#u