pkgs/profpatsch/text-letter: generate PDF with monospace font
This uses the system monospace font to generate a PDF. Before the
printer would be sent a .txt, and it varies from printer to printer
how plain text files are typeset (if at all). Now it only depends on
the system monospace font, which is SourceCode Pro in my case (obvious
TODO is to fix that sometime, or make it configurable).