tests: Fix running background command for PSI test
Recently, the implementation behind Machine.execute() and thus also
Machine.succeed() has been changed[1] to pipe all the command's output
into base64 on the guest machine.

Unfortunately this means that base64 is blocking until stdout is closed,
which in turn means that we now need to make sure that whenever we run a
program in background via "&" we also need to make sure to close stdout.

In the PSI test, we're doing this by simply redirecting the output to
stderr.

[1]: https://github.com/NixOS/nixpkgs/pull/142747

Signed-off-by: aszlig <aszlig@nix.build>