machines/gunnr: Fix evaluation of kernelPackages
The boot.kernelPackages option is already defined in
devhell/profiles/base.nix to be linuxPackages_latest with the same
priority (100) as in the machine configuration of "gunnr".

While generally I leave it for respective owners of the machine
configuration to fix, this however breaks evaluation of all of our
tests.

Given that "latestCompatibleLinuxPackages" is used on a ZFS system, I
assume that this is what the author wanted with higher priority back in
678754b4a5ac5a468e2223e1b701b8e9e2375c30.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @devhell