flameshot: Rebase patch against version 11.0.0
This is a major upstream release which deprecates a few of our options
but also made our patch a little longer because the tools we remove are
now defined in multiple places.

The patch itself is pretty much doing the same things as before we do
have one thing that is different, which is that we now create the
savePath directory if it doesn't exist. I do know that this doesn't
necessary need to be a directory but in the way I'm using flameshot it
is, so I don't care ;-)

While at it, I also disabled the checkForUpdates option, because this
has annoyed me for some time and I really don't care if the version is
the latest for a screenshot utility.

Signed-off-by: aszlig <aszlig@nix.build>