machines/sternenseemann: use upower over low-battery
I had some weird issues with the low-battery udev rule, mainly it not
triggering when it should. Usually, the event would only get processed
when the battery changed state, e.g. from Discharging to Charging.
Consequently, the laptop would hibernate when you'd save it from running
out of battery by plugging it in, but, if you forgot, it'd be content to
run out of battery.

I'll try upower instead now which is the “normal” solution used by the
major desktop environments. It's has some extra complexity, as it also
provides a d-bus API for other applications to use, but we'll see how it
goes.