profiles/packages: Add packages
I'm particularly interested in `gurk-rs` as it's a Signal client that
runs in the terminal. It's in early development but it looks already
amazing and it doesn't rely on the Java library.

`termusic` is also a nice music player written in Rust that I've come to
enjoy when MPD doesn't make sense.

Lastly, `writedisk` is just mad awesome. It can write ISOs of all sorts
to USB stick, and it even knows how to deal with Windows ISOs. It writes
those ISOs really fast as well, so I'm guessing it somehow measures the
ideal block size. Either way, this is much more convenient than having
to invoke `dd` every time I have to write an ISO. Funny enough, this too
is a Rust application.