dnyarri: Add default printer
This is one of the few things that I originally added manually, simply
because usually printing is not something I do regurarily.

Since I had to re-do the CUPS config a few weeks ago, I decided to
instead switch to ensurePrinters, simply because it makes it easier to
just remove all the CUPS state and get to a known working configuration
in the event that something breaks.

Signed-off-by: aszlig <aszlig@nix.build>