machines/dnyarri: Use async discard for root FS
Another thing that was lurking around in configuration.nix, so it has
been battle-tested for weeks. Given that all the layers from the
hardware up to the LUKS container with the filesystem support discard,
it does make sense to enable it.

The disadvantage of using discard with LUKS is that attackers can now
gain information about the file system in use. However, this is already
public knowledge so given that discard increases performance, I decided
to enable it despite some warnings[1] out there.

[1]: https://asalor.blogspot.com/2011/08/trim-dm-crypt-problems.html

Signed-off-by: aszlig <aszlig@nix.build>