machines/tishtushi: Switch to new SSD
Previously, the machine had an internal SSD with a size of around 20 GB,
which until it died in 2018[1] was used as the /nix/store.

With dnyarri getting a hardware upgrade[2], the SSD that was used back
then as a bcache became obsolete.

Given that the spinning rust inside Tishtushi is slow as hell and also
way smaller (320 GB) than the former bcache SSD (512 GB), I decided to
just replace the spinning rust with the SSD, which should at least make
I/O operations bearable while falling asleep on CPU-bound operations.

Since the NixOS LUKS setup now also propagates passphrases to other
devices, we also no longer need the vault device because the reason it
existed was to avoid being prompted multiple times for the passphrase.

[1]: 54c99271f7570c32a4215a097e577272cd4210b2
[2]: be0fb40a12b5a9301509ad45fda1eda11971fa8e

Signed-off-by: aszlig <aszlig@nix.build>