machines/tishtushi: Add xxhash to available mods
I created the BTRFS module with a few different options than I did
initially, one of them is that I changed the checksum algorithm to
XXHASH.

Unfortunately, the xxhash_generic module providing the algorithm is not
available during stage 1, so mounting the volume would fail.

In the local configuration.nix I already added it to
availableKernelModules, but when I moved the config over to Vuizvui in
e04c7e04f2fe63cece7a2f2064212bc055cdd035, I forgot to add it.

Signed-off-by: aszlig <aszlig@nix.build>