pkgs/rustfmt: Fix patching in default config
The patch file was relative to the rustfmt crate within the rustc source
tree, so I added "-d src/tools/rustfmt" as patchFlags to apply the patch
against that subdirectory.

Unfortunately, patchFlags also applies to other patches as well and I
didn't account for that. With current rustfmt from rustc 1.59.0 we have
a patch in nixpkgs, which now fails to apply because of this:

  applying patch /nix/store/...-rustfmt-fix-self-tests.patch
  can't find file to patch at input line 3
  Perhaps you used the wrong -p or --strip option?
  The text leading up to this was:
  --------------------------
  |--- a/src/tools/rustfmt/src/ignore_path.rs
  |+++ b/src/tools/rustfmt/src/ignore_path.rs
  --------------------------
  File to patch:
  Skip this patch? [y]
  Skipping patch.
  1 out of 1 hunk ignored

To fix this, I changed the paths in config.patch to be relative to the
rustc tree rather than just the rustfmt crate, so the patch should now
apply regardless of whether there are other patches.

Signed-off-by: aszlig <aszlig@nix.build>