tests/gnupg: Work around possible buffering issue
Recently the test has started to stall, so I investigated and found the
commit that introduced this in nixpkgs[1]. The commit in question
changes the command to be passed to Machine.execute from a list in a
subshell environment (like "(some commands)") to using a subshell
directly via "sh -c". This is allegedly to avoid shell injection, but
what matters in our case I guess is that it forks into another shell.

Unfortunately, I did not find out *exactly* why this happens but it
seems to affect buffering in such a way that the command never
terminates for some reason.

Since I don't have a lot of time to investigate further, I just worked
around the issue by suppressing stdout output from the expect scripts we
run. This is not nice but since the test stall is irrelevant to our
gnupg module, it doesn't make sense to block the test for reasons out of
the scope of this test.

[1]: https://github.com/NixOS/nixpkgs/commit/dbc95f15b8dad5224cbb6a52df9

Signed-off-by: aszlig <aszlig@nix.build>