profiles/workstation: Allow unraw via sysrq
I just was forced to reboot my workstation while working on a program
that sets the VT into raw mode but finding out that the unraw key now
doesn't work anymore. This used to be the case but with the sysctl
include that comes with systemd-coredump, the kernel.sysrq value
defaults to only allowing sync (16).

Signed-off-by: aszlig <aszlig@nix.build>