pkgs/vim: Fix elflord color theme
I know, I know, people don't like the color scheme I use, but I got so
used to it and with the new version generated via ColorTemplates it
messes with me being used to elflord with 16 terminal colors instead of
the GUI colors.

So this patches elflord in such a way that it's consistent with what I
want but still uses the new ColorTemplates version, which for example
comes with nicer colors for vimdiff.

Signed-off-by: aszlig <aszlig@nix.build>