profiles/base: Use XKB config for console keymap
This is mainly for having a consistent and central way to define our
keybord config. The main reason for doing so is because of Slylandro,
which comes with a keyboard that has a caps lock key atop the left shift
key and it drives me mad.

Since I'd like to keep things DRY, let's just use XKB for everything
keymap-related.

Signed-off-by: aszlig <aszlig@nix.build>