packages: Replace jwhois by whois
The upstream project seems to be no longer maintained[1] and the last
release from 2005 contains a whole bunch of outdated NIC information
so that it became pretty annoying to use.

Back then jwhois was the only whois client packaged in nixpkgs, so this
might be the reason why it ended up in my setup in the first place.

The "whois" package on the other hand seems to be actively maintained
and works well for the domains I care about.

[1]: https://github.com/jonasob/jwhois/issues/32

Signed-off-by: aszlig <aszlig@nix.build>