machines + modules: Fix Nix option definitions
Most of the options in nix.conf are now exposed as a submodule with a
freeform type and since that change[1] got introduced, we get a bunch of
warnings during machine evaluation:

  trace: warning: The option `nix.useSandbox' defined in `...' has been renamed to `nix.settings.sandbox'.
  trace: warning: The option `nix.maxJobs' defined in `...' has been renamed to `nix.settings.max-jobs'.
  trace: warning: The option `nix.buildCores' defined in `...' has been renamed to `nix.settings.cores'.

To shut them up, I went through all machines and modules and renamed the
remaining options that were not renamed back then when @devhell did some
renames in a0297bf921399c3243dcca99626d8697f0735abe.

This was done by looking through the output of:

  $ git grep -A 10 '\<nix\(\.\| *=\)' machines modules

After that I tested the contents of the nix.conf of all the machines
against the changes this commit introduced via the following command:

  $ nix-build --no-out-link -E '
      with import <nixpkgs/lib>;
      map (m: m.eval.config.environment.etc."nix/nix.conf".source)
          (collect (m: m ? eval) (import ./machines))
    '

I've sorted the resulting nix.conf files and diffed on that result and
the only difference that showed up was the following:

   allowed-users = *
  -auto-optimise-store = false
   auto-optimise-store = true
   builders-use-substitutes = true
   cores = 0

This is because the previous way to generate the config was by
concatenating strings whereas the new way works on an attribute set, so
we get deduplication by design.

[1]: https://github.com/NixOS/nixpkgs/pull/139075

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @devhell
Cc: @Profpatsch
Cc: @sternenseemann