pkgs/vim: Remove hex and color names highlighting
We only want to highlight #aabbcc-style hex colors and not every damn
hex value that happens to be 3 or 4 bytes, which is very annoying when
you're doing a lot of bit manipulation.

Additionally, I currently have a code base where "gold" is something a
player can have, so one doesn't need much imagination on how irritating
that is.

Signed-off-by: aszlig <aszlig@nix.build>