profiles/workstation: Use autoLogin on LUKS
After countless boots with unlocking my LUKS containers and thereafter
logging in via lightdm I started to wonder whether it really makes sense
to do the latter.

All of my workstations are single-user, I usually don't switch my
session type (and even if, then via system configuration) and if someone
manages to unlock my LUKS containers, it's really trivial to circumvent
the user authentication.

So if I'm not forgetting about something big[TM], the only purpose this
additional auth serves is me being annoyed for no reason, so let's
disable it if the configuration indicates that LUKS is used.

Signed-off-by: aszlig <aszlig@nix.build>