pkgs/vim: Add quick-scope plugin
Found this thanks to @fpletz starring the repository and since I
constantly use f/F and t/T when navigating, I decided to give it a try.

Usually when trying things, I'd just add the Vim version with the plugin
to my local Nix profile, but since I'm already irritated when other
things than the cursor are highlighted in the current line, I decided to
do something completely irrational and just add it to all my machines.

Signed-off-by: aszlig <aszlig@nix.build>