slylandro: Add light command
While I do use redshift to decrease the brightness at night, I also tend
to sometimes use the laptop when barely awake. Having a convenient CLI
makes it easier to script, eg. "mpv somefile; light -S 0".

The reason why I did not add this to my generic workstation profile is
because my workstations are usually placed in a way that I need to be
awake enough to approach them.

Signed-off-by: aszlig <aszlig@nix.build>