slylandro: Switch to PipeWire
I've been testing PipeWire on this machine for quite a while and it
works pretty well so far. My goal is to switch all my machines to it,
but since Dnyarri currently has a pretty complicated audio setup I
decided to not make it the default for the workstation profile yet.

Signed-off-by: aszlig <aszlig@nix.build>