channel: Fix exposing meta.isHydraChannel
The isHydraChannel meta attribute is needed in order to tell Hydra that
the derivation in question should be a channel tarball. However in
Nixpkgs the meta attribute is not used, so checkMeta doesn't recognise
it as a valid attribute which leads to an evaluation error.

Recently[1] a commit got merged, which enables shallow type checking for
meta attributes by default. This led to an evaluation error for our
Hydra machine channels for the reason mentioned above.

I opted to work around that issue by adding meta.isHydraChannel after
mkDerivation, because adding isHydraChannel as a valid meta attribute to
Nixpkgs doesn't feel right to me since it's only relevant for Hydra and
its apparently deprecated[2] channel feature.

[1]: https://github.com/NixOS/nixpkgs/commit/6762de9a28e248f46bd0810e03c
[2]: https://github.com/NixOS/hydra/blob/53335323ae79ca1a42643f58e520b376898ce641/doc/manual/src/jobs.md#meta-fields

Signed-off-by: aszlig <aszlig@nix.build>