profiles/packages: Add timewarrior and wtf
I've been using taskwarrior a lot again lately, and this time I'd be
interested in seeing how much time I actually spend on certain tasks, so
I'm adding timewarrior to the mix. Unfortunately the hook requires that
python3 is installed.

WTF is a neat terminal dashboard that show information that I deem
useful and would like to keep an eye on without having to manually check
for stuff.