profiles/services: remove redshift and geoclue2
Switching the redshift package to gammastep didn't work. Somehow there's
a gammastep.service for the user, but I'm still not sure where that
comes from. It also doesn't start automatically. I'm not sure if geoclue
is actually working, so I'm removing everything for now and will look at
it again at a later date.