profiles/packages: remove redshift again
So, seems like the redshift service would only work if there was such a
thing as `graphical-session.target`, which doesn't exist. The mysterious
appearance of `gammastep.service` also won't work because of the same
reason. `gammastep.service` should probably be wanted by
`graphical.target`, which exists. I'm going to do this differently now
and execute gammastep directly in sway.