aacolorize: Convert to Python 3
Python 2 was sunset a long time ago and recently got marked as insecure,
so evaluation fails.

Since this program is very little in size, I decided to quickly switch
to Python 3 in a low-effort way. This means that I didn't switch to
using pathlib or using context managers for files, only the things
necessary so that it still works.

Signed-off-by: aszlig <aszlig@nix.build>