workstation/packages: Remove Python 2 stuff
All the packages (hetzner, pep8 and polib) have Python 3 support but
I've used the Python 2 versions of them. Since Python 2 is failing
evaluation because it's marked as insecure, let's switch to the Python 3
versions.

I also removed the the Python 2 interpreter, especially since I haven't
been using it for quite a while.

Signed-off-by: aszlig <aszlig@nix.build>