profiles/base: Explicitly set locale categories
The default locale is en_US.UTF-8, but this also comes with a 12 hour
clock and a few other nuisances such as imperial units. So while I do
prefer a German locale I don't want things to be in the German
*language*, so setting something like de_DE.UTF-8 for LC_TIME doesn't
fly very well, since it could result in something like "Freitag, 20.
Januar".

To cope with this, I used the C locale to make sure that we get a 24h
clock and english week/month names. Similarily I do not want floats to
be formatted with a comma.

Signed-off-by: aszlig <aszlig@nix.build>