manual: use historical nixpkgs revision to obtain needed xls files
https://github.com/NixOS/nixpkgs/pull/212289 removes the xls files we
relied on in favor of a new solution that goes from options.json to
docbook using a python script. This commit is a band-aid approach of
pulling in the necessary files from an old nixpkgs revision to fix the
build of the manual until we can solve it properly.

Since the intention of upstream is to get rid of docbook for good, it
probably makes sense to port our module documentation to markdown
already when porting our manual build process to the new toolchain.