machines/sternenseemann/wolfgang: install msr-tools
wrmsr can be used to disable an erroneous CPU throttling that is
happening with my new battery. Since that needs to be reset after every
power state switch, having the tool in PATH is a great help.

Eventually I'll fix this in a more permanent fashion via a systemd
oneshot service.