vim/markdown: Remove dircolors syntax
The syntax file in question queries features from the current terminal
and will bring editing Markdown files to a crawl, freezing Vim for more
than ten seconds.

I haven't encountered a dircolors(1) file within a Markdown fenced code
block yet and even if this should be the case someday it's not worth the
high runtime cost.

Signed-off-by: aszlig <aszlig@nix.build>