slylandro: Remove hidpi specific settings for now
This reverts commit c88fd9eaa12c8a3c06502b09c2056d3c91421952.

The hardware.video.hidpi.enable option was removed a while ago[1]
because it's not clear what a single boolean option should mean, so it
doesn't make sense anymore to make any of our options depend on it.

Forthermore, I'm experimenting with different Wayland compositors at the
moment and most of the stuff that I did here is for Xorg. I expect most
of the stuff to be gone when I'm settled with a Wayland setup that works
for me.

[1]: https://github.com/NixOS/nixpkgs/pull/222689

Signed-off-by: aszlig <aszlig@nix.build>