psi: Update to latest Git master
There is no particular reason for this other than me "feeling" that my
version became somewhat old. From looking at the upstream diffs the
changes are mostly fixes.

However, I also updated the QDarkStyle plugin, which does have quite a
few changes and the overall theming now looks lighter than before but
still not in a way that feels like an annoyance. We'll see and revert if
necessary.

Signed-off-by: aszlig <aszlig@nix.build>