hardware/pulse1502: Fix build against kernel 6.3
The nvme-suspend-quirk patch doesn't apply anymore with kernel 6.3
because kernel 6.3 added a new quirk
(NVME_QUIRK_NO_SECONDARY_TEMP_THRESH) to the nvme_quirks enum.

To fix this, I just edited the patch to refer to the newly added quirk
and increase the bitflag by one.

With this, the patch applies, but I haven't yet built the kernel nor did
I actually test this on the target machine.

Signed-off-by: aszlig <aszlig@nix.build>