Revert "firefox: Make FOUC less annoying"
This reverts commit e868bfeca5501c7c02f1fc1013ea6cc7e79af043.

Unfortunately, this causes some sites to be outright unreadable since
they seem to rely on a certain set of default colors.

Given that the FOUC is already pretty minimal in Firefox, I think it's
really not worth breaking a bunch of sites just to get rid of the
flashing, at least until we found a better way to prevent it entirely.

Signed-off-by: aszlig <aszlig@nix.build>