profiles: Revert "Use Nix stable for nixos-option"
This reverts commit b9a7b3f0710ed828c9ec7562a6e641810fadd61b.

Our override is no longer needed and it actually doesn't build anymore
since even current stable Nix (2.17.0) is too new for nixos-option. This
is why in nixpkgs, there is a specific override for nixos-option which
uses Nix 2.15 instead.

For us this essentially means that we can drop the override, since
either someone fixes nixos-option to work with newer Nix versions or we
get a specific override that doesn't interfere with our (as in Vuizvui)
override of the "nix" attribute.

Signed-off-by: aszlig <aszlig@nix.build>