machines: Remove tyree and brawndo
The former no longer exists and the latter has a new hardware and it's
self-managed by the owner now (which is always the best IMHO).

With both machines gone, we no longer need the "managed" profile and the
BFQ module that was used for the old hardware of Brawndo was no longer
used anywhere else. Same with the T100HA module, since Tyree was an ASUS
T100HA and apart from the module being probably very much outdated we
don't have another such hardware in Vuizvui.

Signed-off-by: aszlig <aszlig@nix.build>