Fix eval error due to removed module
In 75ea723caea05056810dad9160bc93e571c77985, I removed a bunch of
modules in a series of fixups but forgot to do a final eval check.

This broke evaluation for all the machines and tests because the
module-list.nix file is not only needed for all the machines but also
for figuring out which tests we need to run for which machine config.

Signed-off-by: aszlig <aszlig@nix.build>