machines: Remove "mailserver"
This one never was in use and it was WIP code to do a more fleshed out
mailserver configuration that should match an Ansible deployment of a
mailserver we had back then at OpenLab.

The machine was never in use (which is apparent from its configuration)
and I even *added* it to Vuizvui from "LaberNix" (Vuizvui's predecessor)
in 915e56fb4453b0701a423b0c96fb145318162ffd probably just for the sake
of completeness.

Signed-off-by: aszlig <aszlig@nix.build>