core/tests: Fix checks for {doku,media}wiki
The module options have been restructured in a way that the attribute
set for defining the actual sites is now in a dedicated "sites" option
instead of eg. services.mediawiki directly, so the checks of whether we
should include this for testing always return true since
services.mediawiki is *always* a non-empty attribute set.

Signed-off-by: aszlig <aszlig@nix.build>