firefox: Fix evaluation with newer nixpkgs
A recent nixpkgs change[1] added an explicit requireSigning attribute,
which in turn sets MOZ_REQUIRE_SIGNING to a zero-width string.

The change also changed the check in the Firefox wrapper to check
whether the browser derivation exposes a requireSigning attribute, which
it doesn't.

Since I'm busy with other things and just want my machine configuration
to evaluate, I quickly worked around this by re-adding the
requireSigning (and allowAddonSideload for that matter) to the
derivation attributes on our side.

[1]: https://github.com/NixOS/nixpkgs/commit/c4c81ac8a219e8725dae514f49c

Signed-off-by: aszlig <aszlig@nix.build>