firefox: Enable Global Privacy Control
This was introduced in Firefox 120.0 and essentially implements the
following specification:

https://privacycg.github.io/gpc-spec/

Similar to Do-Not-Track, it implements a header (Sec-GPC) that expresses
the persons preference to the site of whether the site may sell/share
the browsers interaction.

Since Do-Not-Track pretty much is ignored on most sites that formerly
adhered to it, I expect a similar fate for GPC, but maybe this time
things work out differently (crossing fingers).

Signed-off-by: aszlig <aszlig@nix.build>