firefox: Update ff2mpv to version 5.1.0
Most changes are about installing the native messaging app and packaging
on various distributions and OS, additionally it also tries to support
more browsers.

In addition to that, there is now a way to set mpv arguments via the
plugin's new options page. We however want to set this via Nix, so I'm
completely ignoring that feature and didn't include it in our ff2mpv.py.

Signed-off-by: aszlig <aszlig@nix.build>