vim: Don't rustfmt without an explicit config
This is getting annoying when contributing to external projects that
don't have a rustfmt.toml, because I constantly forget to disable it for
a buffer and then get a massive diff that I need to revert.

My own projects all have a rustfmt.toml anyway, so there really is no
reason to do it by default nor have a custom rustfmt config in Vuizvui.

Signed-off-by: aszlig <aszlig@nix.build>