tests/flameshot: Remove icewm background image
The icewm start command has been changed[1] to be "icewm-session" a
while ago, which not only starts the main icewm program but a bunch of
related services, one of them being icewmbg.

We're doing an OCR of the flameshot UI, which is layered over the screen
with partial transparency, which in turn makes the OCR result pretty
unreliable since now we have a background image.

Since we're not testing icewm but flameshot, I decided to just use the
"icewm" command we've used before, since we're not interested in all the
auxiliary programs that form a full "icewm-session".

This fixes the test since with the background image gone, OCR is new
able to correctly pick up the text from the flameshot UI.

[1]: https://github.com/NixOS/nixpkgs/commit/73cb33678527ddc323c7fcef3d17ac85603e5b02

Signed-off-by: aszlig <aszlig@nix.build>