profiles/base: Add termdown to systemPackages
I used to use coreutils "sleep" for simple countdown timers and given
that I get an urgency hint whenever a command completes, this made a
pretty good stopwatch.

However, with "sleep" it's a little bit tricky (eg. via procfs) to find
out the elapsed time, which is something I occasionally need.

Termdown does that and also displays the countdown in a more fancy way
via Figlet, so it's exactly what I want for this purpose.

Signed-off-by: aszlig <aszlig@nix.build>