pkgs/aszlig/gpodder: Fix evaluation error
Fixes the following evaluation error:

  error: Python version mismatch in 'gpodder-git-2023-07-24':

  The Python derivation 'gpodder-git-2023-07-24' depends on a Python derivation
  named 'python3.12-dbus-python-1.3.2', but the two derivations use different versions
  of Python:

        'gpodder-git-2023-07-24' uses /nix/store/747rhv8hh2w8kqwyni7jcfqzxpldxxcv-python3-3.11.9
  'python3.12-dbus-python-1.3.2' uses /nix/store/l014xp1qxdl6gim3zc0jv3mpxhbp346s-python3-3.12.4

The reason why this error happens is because we override gpodder with
our own version directly from Git and supply additional dependencies
from python3Packages, which is always the latest Python version (3.12 in
this gase). This obviously doesn't mix well with the version used in
nixpkgs, which is still Python 3.11.

Signed-off-by: aszlig <aszlig@nix.build>