modules/tests: don't try running broken wg-quick test for linux 5.4
Unfortunately the version the current default/LTS kernel was never
updated in nixos/tests/wireguard ever since the introduction of the
multi version test feature. https://github.com/NixOS/nixpkgs/pull/338632
will tie it to the default version again after which we can probably
reintroduce a logic similar to this.

Testing the 5.4 version is a bit pointless at the moment. Everyone in
vuizvui seems to be using either the current default version of
linux (6.6) or the latest version, so the latest kernel version is
always “closer” than 5.4.