core/tests: Split ZFS stable tests into series_2_X
Recentry the "stable" attribute in the ZFS NixOS test has been
renamed[1] to "series_2_2", so we now get an evaluation error because of
the missing "stable" attribute.

Quoting the commit in question:

> zfs_2_2: Rename from zfsStable
>
> The `zfs` alias already has equivalent semantics. Instead, make this
> like zfs_2_1 so folks who want to pin a specific release series can do
> so easily and clearly to have more control over when more substantial
> updates occur.
>
> Rename all tests to match the pkg attr they are testing.

To fix this, I not only renamed the attribute on our side but also added
the series_2_1 attribute to our heavily outdated core/tests.nix module.

[1]: https://github.com/NixOS/nixpkgs/commit/dcff4f831836e249db5e808aeac

Signed-off-by: aszlig <aszlig@nix.build>
Reported-by: @devhell