machines/aszlig: Remove autodefrag mount option
Recently I had to work with an SQLite database with several gigabytes in
size and found that the btrfs-cleaner kthread was pretty busy with
btrfs_run_defrag_inodes() continuously running on the inode of the
SQLite database, which resulted in a lot more writes than what I'm
comfortable with on NVMes.

Since disabling autodefrag on a per-file basis is only possible by
losing certain other features (eg. CoW), I decided to disable it
altogether because defragmentation can be triggered from userspace. I'll
keep monitoring fragmentation stats from now on and decide whether it
really was worth enabling autodefrag for my workloads and if not,
whether a user-space autodefrag implementation would be a better option.

Signed-off-by: aszlig <aszlig@nix.build>