flameshot: Remove invalid checkForUpdates setting
The nixpkgs package includes the DISABLE_UPDATE_CHECKER cmake option to
disable the update checker along with its settings.

This causes the flameshot test to fail with a dialog about the settings
being wrong. Since flameshot has implemented configuration checks for
quite a while, we might want to check this at build time at some point
in the future.

Signed-off-by: aszlig <aszlig@nix.build>